[RemoveDI][Polly] Use iterators instead of instruction pointers to SetInsertPoint
As part of the effort to transition to using Debug Records instead
of Debug intrinsics, some API/argument changes are necessary to
achieve the desired behavior from Debug Records. This particular fix
involves passing iterators instead of instruction pointers to the
SetInsertPoint function. While this is crucial in certain areas,
it may be more than needed in others, but it does not cause any harm.

Change-Id: I26872d176dc1c90fba1a85b30854ad37be9fc4c7